setup

+++ src/main/java/com/bihuniak/educator/basic/MyFirstRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MyFirstRepository extends CrudRepository<Car, Integer> {
+public interface MyFirstRepository extends CrudRepository<Car, Long> {
 }
